Back test by vTri

+++ backend/src/main/java/com/backend/backend/controllers/RecipeController.java
@@ -73,6 +73,9 @@ public class RecipeController {
         System.out.println("Deleting Recipe");
         try {
             int recipeId = Integer.parseInt(rid);
+            if (!recipeRepository.existsByRid(recipeId)) {
+                throw new Exception("Recipe Id non-existent");
+            }
             recipeRepository.deleteById(recipeId);
             response.setStatus(200); // 200 = OK
             return true;
@@ -89,6 +92,7 @@ public class RecipeController {
     public List<Recipe> getAllRecipes(HttpServletResponse response) {
         System.out.println("Getting all Recipes");
         try {
+            response.setStatus(200);
             // get all users from database
             List<Recipe> recipeList = recipeRepository.findAll();
             return recipeList;
@@ -107,6 +111,7 @@ public class RecipeController {
             int recipeId = Integer.parseInt(rid);
             Optional<Recipe> recipeOptional = recipeRepository.findById(recipeId);
             if (recipeOptional.isPresent()) {
+                response.setStatus(200);
                 return recipeOptional.get();
             } else {
                 response.setStatus(404); // 404 = not found
